Added an autoloader for classes in the library directory
What this does is if a class hasn't been preloaded via an include/require, it will attempt to find the class in the library classes section.

The filenaming is strict:
 - it must be lowercase
 - it must end with .php
 - the class must be the same name as the file (example new myclass would load myclass.php)

Added an init() to library, this will be called on the load of the library class and it scans the database directory for support database types.
To add a database the file must be database_xxx.php where xxx is the database type (example database_mysql.php)
The class within the database must match the file name (class database_mysql extends database)

Note: you can still use
include "myclass.php";
new myclass;

To create a class pointer, this is just for the library specific classes